Basic functions
I've fix a little bit the structure of the files and codes in them. I've
added a new library to ini a vertical tabs. I've written some Angularjs
structure code, where I send a request to a local point and retrieve
JSON information about the house and populate the UI. I've added some
function for better navigation thru the UI.